Added path with trees, Camera is a little screwed up until you walk
I added trees and grass to the scene just to see how it would look. I
also for some reason dont know why the camera is the way it is but it
fixes itself when the player starts moving, but the hook shot does not
shoot in the right place anymore.